+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 __pycache__
 .ipynb_checkpoints
+.vscode
