+++ backend/routers/__init__.py
@@ -1,0 +1,5 @@
+from .algolist import router
+from .files import router
+from .outputs import router
+from .run import router
+from .params import router
